refactor: add export support removing i18n and api route
