+++ .env
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_PARTNERS_DATA_URL=https://admin.dclstudios.org
