+++ next.config.js
@@ -5,10 +5,6 @@ const withTranspiledModules = require('next-transpile-modules')(['decentraland-u
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
-  i18n: {
-    locales: ['en-US'],
-    defaultLocale: 'en-US',
-  },
 }
 
 module.exports = withTranspiledModules(nextConfig)
