+++ src/clients/Partners.ts
@@ -1,14 +1,29 @@
 import { VerifiedPartner } from '../interfaces/VerifiedPartner'
 
-export default class Partners {
-  static Url = process.env.NEXT_PUBLIC_BASE_URL || ''
+const VERIFIED_PARTNERS_URL = `${process.env.NEXT_PUBLIC_PARTNERS_DATA_URL}/items/profile`
 
-  static async fetch<T>(path: string): Promise<T> {
-    const response = await fetch(`${this.Url}/api${path}`)
-    return (await response.json()) as T
-  }
+export default class Partners {
+  static Url = process.env.NEXT_PUBLIC_PARTNERS_DATA_URL || ''
 
   static async get() {
-    return this.fetch<VerifiedPartner[]>('/partners')
+    let isFinished = false
+    let offset = 0
+    const partners: VerifiedPartner[] = []
+    while (!isFinished) {
+      try {
+        const response = await fetch(`${VERIFIED_PARTNERS_URL}?offset=${offset}`)
+        const data = (await response.json()).data as VerifiedPartner[]
+        if (data.length === 0) {
+          isFinished = true
+        } else {
+          offset += data.length
+        }
+        partners.push(...data)
+      } catch (error) {
+        console.log('error', error)
+      }
+    }
+
+    return partners
   }
 }
